First phase of the scaping of dataset background
A user would want to introduce a new dataset into the experiment environment.
The first phase introduces the channel that will launch the scraping of the
openneuro website and insert the data to the DB (async). The task requires the
doi and version of the dataset.

In the second phase, the scraping task will be completed with downloading the
artifacts and plugging into a API endpoint.